CoreGraphics: Slightly shrink excluded clip regions in order to draw transformed opaque components without border artefacts
Previously, drawing an opaque, scaled component with CoreGraphics could
lead to visible artefacts around the edge of the component.

When drawing the parent of an opaque component, the area covered by the
opaque component is excluded from the clip region. If the clip region is
non-integral when transformed into device space, anti-aliasing will be
applied on the edges of the clip region. Similarly, when drawing the
opaque component itself, anti-aliasing will be applied at the edges of
the component. When the two drawings are superimposed, the foreground
anti-aliased pixels will be blended with the background anti-aliased
pixels, leading to a noticeable border around the component. Ideally,
only the foreground anti-aliasing should be applied, and the background
should not be anti-aliased around its edges.

+++ modules/juce_graphics/native/juce_CoreGraphicsContext_mac.h
@@ -144,7 +144,7 @@ private:
     void flip() const;
     void applyTransform (const AffineTransform&) const;
     void drawImage (const Image&, const AffineTransform&, bool fillEntireClipAsTiles);
-    bool clipToRectangleListWithoutTest (const RectangleList<int>&);
+    bool clipToRectangleListWithoutTest (const RectangleList<float>&);
     void fillCGRect (const CGRect&, bool replaceExistingContents);
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (CoreGraphicsContext)
